* dev:
  [Fix] 로그인 관련 에러사항 수정
  [Fix] 요청 중복 방지 코드 추가
  [Ffx] 빌드했을 때 토큰 못불러오는 부분 및 모달 토큰삭제 제거
  [Fix] 공고 등록 후 페이지 이동 시 이탈 방지 창 출력 방지
  [Fix] 채용 링크 http 없이도 외부 링크로 연결
  [Fix] 2차 직무 필수로 변경, 각 2차 직무 리스트에 "없음" 추가
  [Fix] 선택 사항들 모든 칸 삭제 방지
  [Feat] modal 테스트를 위한 임시 커밋
  [Fix] 공고 등록 시 etc_arr, preferred_etc_arr 및 place에 공백 추가되지 않게 변경
  [Fix] 공장 없을 시 버튼 비활성화 및 안내 문구 추가, 교대 형태 선택 시 문구 수정
  [Fix] 1, 2차 직무 밖 클릭 시 창 꺼지게 변경
  [Fix] 공고 리스트 배경 색상 변경
  [Fix] 공고 등록/수정 시 리스트로 리다이렉트 및 공고 카드 디자인 변경
  [Fix] 직무 카드 하나 남았을 때 직무 삭제 버튼 비활성화
  [Fix] 최대 칸 수에 도달했을 때 입력칸 추가 버튼 없앰

+++ apps/gocho_business/src/pages/jd/[jdId]/edit/index.page.tsx
@@ -1,4 +1,4 @@
-import { ReactElement, useEffect, useState } from "react";
+import { ReactElement, useEffect, useRef, useState } from "react";
 import { SubmitHandler, useFieldArray, useForm } from "react-hook-form";
 import { useRouter } from "next/router";
 import { BiRocket } from "react-icons/bi";
@@ -26,6 +26,7 @@ import { useEndJd } from "@/apis/jd/useEndJd";
 import { jdArrKeyObj } from "@/apis/jd/useJdArr/type";
 
 import { JD_UPLOAD_MESSAGE_OBJ } from "@/pages/jd/upload/constant";
+import { INTERNAL_URL } from "@/constants/url";
 import { HeaderPart } from "./part/headerPart";
 import { BasicInfoPart } from "./part/basicInfoPart";
 import { PositionHeaderPart } from "./part/positionHeaderPart";
@@ -40,6 +41,9 @@ import { cssObj } from "./style";
 
 const JdEditPage: NextPageWithLayout = () => {
   const [isCardOpenArr, setIsCardOpenArr] = useState<boolean[]>([false]);
+  const isEditLoading = useRef(false);
+  const isDeleteLoading = useRef(false);
+  const isEndLoading = useRef(false);
 
   const router = useRouter();
   const queryClient = useQueryClient();
@@ -59,7 +63,7 @@ const JdEditPage: NextPageWithLayout = () => {
     control,
     handleSubmit,
     reset,
-    formState: { isDirty, submitCount },
+    formState: { isDirty, isSubmitSuccessful, submitCount },
   } = jdForm;
 
   const { fields, append, remove } = useFieldArray({
@@ -88,6 +92,9 @@ const JdEditPage: NextPageWithLayout = () => {
   const { mutate: endJdMutation } = useEndJd();
 
   const endJdHandler = (id: number) => {
+    if (isEndLoading.current) return;
+    isEndLoading.current = true;
+
     if (window.confirm(JD_EDIT_MESSAGE_OBJ.END)) {
       endJdMutation(
         { jdId: id },
@@ -96,12 +103,19 @@ const JdEditPage: NextPageWithLayout = () => {
             setToast("마감되었습니다");
             queryClient.invalidateQueries(jdArrKeyObj.all);
           },
+
+          onSettled: () => {
+            isEndLoading.current = false;
+          },
         }
       );
     }
   };
 
   const deleteJdHandler = (id: number) => {
+    if (isDeleteLoading.current) return;
+    isDeleteLoading.current = true;
+
     if (window.confirm(JD_EDIT_MESSAGE_OBJ.DELETE)) {
       deleteJdMutation(
         { jdId: id },
@@ -110,12 +124,19 @@ const JdEditPage: NextPageWithLayout = () => {
             setToast("삭제되었습니다");
             queryClient.invalidateQueries(jdArrKeyObj.all);
           },
+
+          onSettled: () => {
+            isDeleteLoading.current = false;
+          },
         }
       );
     }
   };
 
   const jdEditHandler: SubmitHandler<JdFormValues> = (jdObj) => {
+    if (isEditLoading.current) return;
+    isEditLoading.current = true;
+
     jdEditConfirmEvent();
     if (window.confirm(JD_EDIT_MESSAGE_OBJ.EDIT)) {
       editJdMutate(
@@ -131,6 +152,9 @@ const JdEditPage: NextPageWithLayout = () => {
             etc_arr: getFieldArrayValueWithNull(jdObj.etc_arr),
             position_arr: jdObj.position_arr.map((position) => ({
               ...position,
+              conversion_rate: position.conversion_rate ? position.conversion_rate : null,
+              min_year: position.min_year ? position.min_year : null,
+              max_year: position.max_year ? position.max_year : null,
               hire_number: position.hire_number ? position.hire_number : 0,
               task_sub_arr: position.task_sub_arr ? position.task_sub_arr : null,
               task_detail_arr: getFieldArrayValue(position.task_detail_arr),
@@ -144,11 +168,16 @@ const JdEditPage: NextPageWithLayout = () => {
         {
           onSuccess: () => {
             jdEditDoneEvent();
-            setToast("등록되었습니다");
+            router.push(INTERNAL_URL.JD_LIST);
+            setToast("수정되었습니다");
           },
 
           onError: () => {
             alert("에러입니다. 조건을 한번 더 확인하거나 운영자에게 문의해주세요.");
+          },
+
+          onSettled: () => {
+            isEditLoading.current = false;
           },
         }
       );
@@ -201,10 +230,12 @@ const JdEditPage: NextPageWithLayout = () => {
   }, [jdData, reset]);
 
   useEffect(() => {
-    window.onbeforeunload = () => isDirty;
+    const escapeWithoutSubmit = isDirty && !isSubmitSuccessful;
+
+    if (escapeWithoutSubmit) window.onbeforeunload = () => true;
 
     const handleUnload = () => {
-      if (isDirty) {
+      if (escapeWithoutSubmit) {
         jdEditExitEvent();
         if (!window.confirm(JD_UPLOAD_MESSAGE_OBJ.LEAVE)) {
           throw router.events.emit("routeChangeError");
@@ -220,7 +251,7 @@ const JdEditPage: NextPageWithLayout = () => {
       router.events.off("routeChangeStart", handleUnload);
       window.onbeforeunload = () => null;
     };
-  }, [isDirty, router.events]);
+  }, [isDirty, isSubmitSuccessful, router.events]);
 
   useEffect(() => {
     if (submitCount === 0) return;
